Adds test cases for single digit input and passes them Passes all single digit test cases and throws when the number is negative.

+++ include/string_kata.h
@@ -3,6 +3,6 @@
 
 #include <string>
 
-int stringSum (const std::string & inputStr);
+long stringSum (const std::string & inputStr);
 
 #endif
